Refactor ColorPickerView
